test: add comprehensive unit tests for getCandleStartTimestamp
- Verified correct timestamp calculations across various timeframes:
  - Tested basic timeframes (1 minute, 1 hour) with small candle counts.
  - Covered zero and negative candle scenarios for graceful handling.
  - Validated large candle counts and boundary conditions to detect overflow.
- Ensured all supported timeframes produce accurate timestamp offsets.
- Handled invalid timeframe inputs with proper exception assertions.
- Improved robustness and edge case coverage for time-based utilities.

+++ include/Helper.hpp
@@ -466,7 +466,7 @@ blaze::DynamicMatrix< T > sliceCandles(const blaze::DynamicMatrix< T > &candles,
 template < typename T >
 int64_t getNextCandleTimestamp(const blaze::DynamicVector< T > &candle, const Enum::Timeframe &timeframe);
 
-int64_t getCandleStartTimestampBasedOnTimeframe(const Enum::Timeframe &timeframe, int num_candles_to_fetch);
+int64_t getCandleStartTimestampBasedOnTimeframe(const Enum::Timeframe &timeframe, int64_t num_candles_to_fetch);
 
 /**
  * @brief Prepare quantity based on side
